+++ jdaviz/configs/default/plugins/collapse/collapse.py
@@ -1,5 +1,3 @@
-import os
-from pathlib import Path
 import warnings
 
 from astropy.nddata import CCDData
@@ -43,7 +41,6 @@ class Collapse(PluginTemplateMixin, DatasetSelectMixin, SpectralSubsetSelectMixi
     function_selected = Unicode('Sum').tag(sync=True)
     filename = Unicode().tag(sync=True)
     collapsed_spec_available = Bool(False).tag(sync=True)
-    overwrite_warn = Bool(False).tag(sync=True)
     # export_enabled controls whether saving to a file is enabled via the UI.  This
     # is a temporary measure to allow server-installations to disable saving server-side until
     # saving client-side is supported
@@ -143,46 +140,3 @@ class Collapse(PluginTemplateMixin, DatasetSelectMixin, SpectralSubsetSelectMixi
 
     def vue_collapse(self, *args, **kwargs):
         self.collapse(add_data=True)
-
-    def vue_save_as_fits(self, *args):
-        self._save_collapsed_spec_to_fits()
-
-    def vue_overwrite_fits(self, *args):
-        """Attempt to force writing the file if the user confirms the desire to overwrite."""
-        self.overwrite_warn = False
-        self._save_collapsed_spec_to_fits(overwrite=True)
-
-    def _save_collapsed_spec_to_fits(self, overwrite=False, *args):
-
-        if not self.export_enabled:
-            # this should never be triggered since this is intended for UI-disabling and the
-            # UI section is hidden, but would prevent any JS-hacking
-            raise ValueError("Writing out collapsed cube to file is currently disabled")
-
-        # Make sure file does not end up in weird places in standalone mode.
-        path = os.path.dirname(self.filename)
-        if path and not os.path.exists(path):
-            raise ValueError(f"Invalid path={path}")
-        elif (not path or path.startswith("..")) and os.environ.get("JDAVIZ_START_DIR", ""):  # noqa: E501 # pragma: no cover
-            filename = Path(os.environ["JDAVIZ_START_DIR"]) / self.filename
-        else:
-            filename = Path(self.filename).resolve()
-
-        if filename.exists():
-            if overwrite:
-                # Try to delete the file
-                filename.unlink()
-                if filename.exists():
-                    # Warn the user if the file still exists
-                    raise FileExistsError(f"Unable to delete {filename}. Check user permissions.")
-            else:
-                self.overwrite_warn = True
-                return
-
-        filename = str(filename)
-        self.collapsed_spec.write(filename)
-
-        # Let the user know where we saved the file.
-        self.hub.broadcast(SnackbarMessage(
-            f"Collapsed cube saved to {os.path.abspath(filename)}",
-                           sender=self, color="success"))